[nixprofile] call nix.ProfileList api
